0327/14:50 update chat room_id
