--- src/components/MainPage(현재 사용 안함).js
@@ -1,14 +1,0 @@
-import React from "react"
-import SessionCardContainer from "../containers/SessionCardContainer"
-import NavContainer from "../containers/NavContainer"
-
-const MainPage = () => {
-    return (
-        <>
-        <SessionCardContainer />
-        <NavContainer />
-        </>
-    )
-}
-
-export default MainPage;
