+++ src/components/liveSession/LiveSession.js
@@ -154,48 +154,48 @@ const LiveSession = (props) => {
 
     //^ =============================================================
     let rtmChannel;
-    // const [channel, setChannel] = useState();
+    const [channel, setChannel] = useState();
 
-    // const {
-    //     localAudioTrack,
-    //     leave,
-    //     join,
-    //     joinState,
-    //     remoteUsers,
-    //     authority,
-    // } = useAgora(client);
+    const {
+        localAudioTrack,
+        leave,
+        join,
+        joinState,
+        remoteUsers,
+        authority,
+    } = useAgora(client);
 
-    // useEffect(() => {
-    //     rtmChannel = rtmClient.createChannel(props.channelNum);
-    //     join(props.channelNum, null, rtmClient, rtmChannel);
-    //     return () => {
-    //         rtmClient.logout();
-    //         leave();
-    //     }
-    // }, [])
+    useEffect(() => {
+        rtmChannel = rtmClient.createChannel(props.channelNum);
+        join(props.channelNum, null, rtmClient, rtmChannel);
+        return () => {
+            rtmClient.logout();
+            leave();
+        }
+    }, [])
 
-    // const onClick = (choice) => () => {
-    //     if (choice === "join") {
-    //       console.log("join");
-    //       console.log(props.channelNum);
-    //       rtmChannel = rtmClient.createChannel(props.channelNum);
+    const onClick = (choice) => () => {
+        if (choice === "join") {
+          console.log("join");
+          console.log(props.channelNum);
+          rtmChannel = rtmClient.createChannel(props.channelNum);
           
-    //       join(props.channelNum, null, rtmClient, rtmChannel);
+          join(props.channelNum, null, rtmClient, rtmChannel);
     
-    //     } else if (choice === "leave") {
-    //       console.log("leave");
-    //       rtmClient.logout();
-    //       leave();
-    //     }
-    // };
+        } else if (choice === "leave") {
+          console.log("leave");
+          rtmClient.logout();
+          leave();
+        }
+    };
 
-    //^ =============================================================
+    // ^ =============================================================
 
     return (
         <>
         <div style={style.livewrapper}>
             <div style={style.livesession}>
-            {/* <div className="layerfordark"> */}
+            <div className="layerfordark">
                 <div style={style.session_top}>
                     <table style={style.table}>
                         <tr>
@@ -239,12 +239,12 @@ const LiveSession = (props) => {
                     <Grid container justify="center">
                         <QuestionSwiper/>
                     </Grid>
-                    {/* <Grid container justify="center">
+                    <Grid container justify="center">
                         <div style={style.Insertfield}>
                                 <InsertField goListUp = {setListUp} goDark={setDark}/>
 
                         </div>
-                    </Grid> */}
+                    </Grid>
                     <div className="forchat"></div> 
                     <Chat goQueUp={setQueUp} goListUp = {setListUp} goDark={setDark} room={room} windowHeight="1000px" onBack={()=>setRoom(null)}/>
                     <div className="chattingblind"></div>
@@ -260,10 +260,10 @@ const LiveSession = (props) => {
             <Questioning goQueUp = {setQueUp} goDark={setDark}/>
         </div>
         <div style={dark} className="layerfordark"></div>
-        {/* <div className="agora"> */}
+        <div className="agora">
                 
 
-        {/* <div className="button-group">
+        <div className="button-group">
             <button
               id="join"
               type="button"
@@ -288,8 +288,8 @@ const LiveSession = (props) => {
             >
               Leave
             </button>
-          </div> */}
-        {/* <div className="host-player">
+          </div>
+        <div className="host-player">
             {client.uid}
             <PlayerWrapper
                 client={client}
@@ -299,12 +299,12 @@ const LiveSession = (props) => {
                 remoteUsers={remoteUsers}
                 channelNum={props.channelNum}
             />
-        </div> */}
-
-
-        {/* </div> */}
+        </div>
+        </div>
+        </div>
         </>
     )
+
 }
 
 export default LiveSession
